Refactor: merge GoUtil.ts and Misc.ts -> util.ts
