+++ src/local_estimators/voronoi.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { dup } from "../GoUtil";
+import { dup } from "../util";
 
 /**
  * This estimator simply marks territory for whichever color has a
